MotionDebugOverlayをDevOverlayに統合
- DevOverlayにモーション検出セクションを追加
- MotionDebugOverlay.tsxを削除
- フェーズ、経過時間、加速度、停車回数を表示
